+++ lib/fig/configurable/types.rb
@@ -17,7 +17,7 @@ module Fig
     # A coercible bool extended to admit enabled/disabled
     #
     # NOTE: it doesn't work without safe, but I'm unsure why as dry-types doesn't seem to explain itself.
-    #       My hunch is it has to do with the fact that Bool is defined as an ADT of True | False, but might be wronf
+    #       My hunch is it has to do with the fact that Bool is defined as an ADT of True | False, but might be wrong
     Bool = ::Dry::Types.module::Form::Bool.constructor(Coercions.method(:enablement_to_bool)).safe
     class << self
       prepend(Module.new do